Fix: java agent 32-bit pointer to jlong warning
Fix the following compiler warning on 32-bit:

15:07:12 lttng_ust_context.c: In function 'Java_org_lttng_ust_agent_context_LttngContextApi_registerProvider':
15:07:12 lttng_ust_context.c:377:17: warning: cast from pointer to
integer of different size [-Wpointer-to-int-cast]
15:07:12   provider_ref = (jlong) provider;

+++ liblttng-ust-java-agent/jni/common/lttng_ust_context.c
@@ -374,7 +374,7 @@ JNIEXPORT jlong JNICALL Java_org_lttng_ust_agent_context_LttngContextApi_registe
 		goto error_register;
 	}
 
-	provider_ref = (jlong) provider;
+	provider_ref = (jlong) (long) provider;
 	return provider_ref;
 
 	/* Error handling. */
